Add option for adding projects and note card module
Added an option to add additional project arrays, to add new cards to.
Added a module that adds the card to the item box.
Still having indexing issues with the arrays that need to be corrected.
Also need to integrate an option to only display the items relevant to
the selected project on pressing of the project list item.