Change to object structure and clean/reorganise code
Had found that the code was becoming more complex due to dependancies
and confusing to index due to the array structure. Decided to hash out
a new structure, mainly converting to an object model and focus on
implementing the SOLID principles. Have started working through the
essential actions and modules. Will continue to attempt to simplify each
feature.