Add project delete, info popup, today/urgent features
Added an option for project deletion, and corrected the project
selection issues. Also added options for displaying notes due for today,
and urgent notes (with priority ranking of 1). Added a description box
popup feature, that can be closed using a checkbox. Restyled the page,
adding a kode mono font and a lined paper style background. Also added
screen size adaptability for under 1000px, not perfect yet but is
sufficient.